refactoring hw09
